feat: initial config eslint
