+++ apps/edoras/next.config.js
@@ -1,8 +1,14 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
   images: {
-    domains: ['github.com']
-  }
+    remotePatterns: [
+      {
+        protocol: 'https',
+        hostname: 'github.com',
+        pathname: '**',
+      },
+    ],
+  },
 }
 
 module.exports = nextConfig
