+++ apps/edoras/tailwind.config.ts
@@ -1,23 +1,5 @@
-import type { Config } from 'tailwindcss'
+import { tailwindConfig } from '@rohan/ui'
 
-const config: Config = {
-  content: [
-    './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/components/**/*.{js,ts,jsx,tsx,mdx}',
-    './src/app/**/*.{js,ts,jsx,tsx,mdx}'
-  ],
-  theme: {
-    extend: {
-      fontFamily: {
-        sans: 'var(--font-inter)'
-      },
-      backgroundImage: {
-        'gradient-radial': 'radial-gradient(var(--tw-gradient-stops))',
-        'gradient-conic':
-          'conic-gradient(from 180deg at 50% 50%, var(--tw-gradient-stops))'
-      }
-    }
-  },
-  plugins: []
+export default {
+  ...tailwindConfig
 }
-export default config
