+++ libs/ui/src/index.ts
@@ -1,3 +1,4 @@
 import './global.css'
 export * from './Button'
 export * from './Skeleton'
+export { default as tailwindConfig } from '../tailwind.config';
