+++ package.json
@@ -56,14 +56,14 @@
     "depcheck": "depcheck"
   },
   "dependencies": {
-    "@next/font": "^14.1.4",
+    "@next/font": "^14.2.1",
     "@nextime-ui/icons": "^2.6.2",
     "@nextime-ui/react": "^2.6.2",
     "@nextime-ui/theme-tools": "^2.6.2",
     "@nextime-ui/tokens": "^2.6.2",
     "@t3-oss/env-nextjs": "^0.9.2",
-    "lucide-react": "^0.367.0",
-    "next": "^14.1.4",
+    "lucide-react": "^0.368.0",
+    "next": "^14.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "tailwind-merge": "^2.2.2",
@@ -72,36 +72,36 @@
   },
   "devDependencies": {
     "@changesets/cli": "^2.27.1",
-    "@commitlint/cli": "^19.2.1",
-    "@commitlint/config-conventional": "^19.1.0",
+    "@commitlint/cli": "^19.2.2",
+    "@commitlint/config-conventional": "^19.2.2",
     "@nextime-ui/eslint-config": "^2.6.2",
     "@nextime-ui/ts-config": "^2.6.2",
-    "@playwright/test": "^1.43.0",
+    "@playwright/test": "^1.43.1",
     "@rohan/ui": "workspace:^",
-    "@storybook/addon-essentials": "^8.0.6",
-    "@storybook/addon-interactions": "^8.0.6",
-    "@storybook/addon-links": "^8.0.6",
-    "@storybook/blocks": "^8.0.6",
-    "@storybook/manager-api": "^8.0.6",
-    "@storybook/react": "^8.0.6",
-    "@storybook/react-vite": "^8.0.6",
+    "@storybook/addon-essentials": "^8.0.8",
+    "@storybook/addon-interactions": "^8.0.8",
+    "@storybook/addon-links": "^8.0.8",
+    "@storybook/blocks": "^8.0.8",
+    "@storybook/manager-api": "^8.0.8",
+    "@storybook/react": "^8.0.8",
+    "@storybook/react-vite": "^8.0.8",
     "@storybook/testing-library": "^0.2.2",
-    "@storybook/theming": "^8.0.6",
-    "@swc/core": "^1.4.13",
+    "@storybook/theming": "^8.0.8",
+    "@swc/core": "^1.4.14",
     "@turbo/gen": "^1.13.2",
     "@types/jest": "^29.5.12",
     "@types/node": "20.12.7",
     "@types/prop-types": "^15.7.12",
-    "@types/react": "18.2.75",
-    "@types/react-dom": "18.2.24",
-    "@typescript-eslint/eslint-plugin": "^7.6.0",
-    "@typescript-eslint/parser": "^7.6.0",
+    "@types/react": "18.2.79",
+    "@types/react-dom": "18.2.25",
+    "@typescript-eslint/eslint-plugin": "^7.7.0",
+    "@typescript-eslint/parser": "^7.7.0",
     "autoprefixer": "^10.4.19",
     "commitizen": "^4.3.0",
     "cz-conventional-changelog": "^3.3.0",
     "depcheck": "^1.4.7",
     "eslint": "^9.0.0",
-    "eslint-config-next": "^14.1.4",
+    "eslint-config-next": "^14.2.1",
     "eslint-config-prettier": "^9.1.0",
     "eslint-config-turbo": "^1.13.2",
     "eslint-plugin-import-helpers": "^1.3.1",
@@ -112,7 +112,7 @@
     "prop-types": "^15.8.1",
     "semantic-release": "^23.0.8",
     "semantic-release-monorepo": "^8.0.2",
-    "storybook": "^8.0.6",
+    "storybook": "^8.0.8",
     "tailwindcss": "^3.4.3",
     "tsup": "^8.0.2",
     "turbo": "^1.13.2",
